Add support for floating point numbers
Floats are stored internally as a float64 like ints are stored
as int64. All integer arithmatic is available for floats.
There are also two builtin functions to_float and to_int. They
do as their name says. They return the given number to a float
or int.

Currently, arithmatic can only be performed on the same types.
Different types will produce an error hence the addition of the
builtin functions.