Allow functions to be called with more than declared parameters
Functions can take optional arguments. Parameters that are declared
at runtime are bound as usual. Anything extra is assigned to the
local variable "args". Args is an array that contains all parameters
BEYOND the defined. So if a function is defined as func(a, b) and
called as func(1, 2, 3, 4), local variables "a" and "b" will be
defined as "1" and "2" respectively. "args" will be defined as
[3, 4]. If no extra args are given, "args" will be nil. Functions
can check for this by calling "len(args)" which returns 0 for nil.