Implemented hashMerge(map1, map2[, overwrite]); len(nil) == 0
Overwrite is an optional argument which is true by default.
When overwrite is true, if a key in map2 is the same as map1,
the value in map2 will overwrite the value in map1. If overwrite
is false, map2's value is simply ignored.

Also changed len() to return 0 for nil. This is originally for
functions to check if they have any extra arguments.