Added toString(), to_int -> toInt, to_float -> toFloat
Since readline() requires a string, thought it'd be nice to
have a conversion function for types to strings. All types
except arrays, maps, and functions are valid input and will
be converted to a string representation.

Renamed to_int and to_float to match toString and because I
don't actually like underscores in functions names.