Implement escape sequences in strings, allow lexer to make errors
Lexer functions can generate an ILLEGAL token which the parser
will add to its errors.

Strings may contain escape sequences. There are now two types of
strings. Raw strings are enclosed in single quotes and the only
escape sequence processed is "\'" which escapes a single quote.

Normal strings allow the following sequences:
\n
\r
\t
\v
\f
\\
\"
\b

Any other sequence will print the backslash and the character following
it as if nothing happened. The single quote is NOT escaped in a normal
string. Literal newlines are not allowed in normal strings. Meaning, a
double quoted string must be completely on a single line.