Use iota/int for TokenType, align token name with keyword
Strings are wasteful for space. I'm already wasting so much
with a bloated AST and objects I don't need to use more for
strings.

Added a String() method to TokenType to return a string
representation of token.

Build the keyword map automatically (inspired by Go).

Renamed a few tokens to match their keywords. let and always
were def and const.