Changed Error object to Exception, created new Error object
The old error object was a stop-the-world type error. It
functions more like an exception than a simple error. So
now it is. Exceptions are still stop-the-world. They should
be used when the programmer cannot fix a mistake programmatically.
Errors are like normal objects. There are two new builtins
to handle them. isError() will return if a variable is an error
object. errorVal() will return the error message. Printing an
error will also print the error message.

Eventually I'll add a try/catch mechanism for catch exceptions.