Implemented an SCGI server for the interpreter
The interpreter can be started with the -scgi flag
which starts an SCGI server listening on 0.0.0.0:9000.
A web server can send SCGI requests to this connection
and the interpreter will parse and execute a nitrogen
script without starting a separate interpreter.

This version only support a single request at a time.
Concurrent requests are coming soon.