Started bytecode compiler and virtual machine
A lot of the language is implemented. The last
major pieces of core functionality are indexed
expressions and functions. Functions I need
to think about a bit to make sure closures
are possible and normal scoping is preserved.
Indexed expressions should be simple enough,
just need to implement a couple of get_index
and set_index opcodes.

Yes this looks a lot like Python. The cpython
vm has decentish documentation or it can be
found easily. PHP doesn't have a lot of explanation
for the Zend vm and I don't even know what the
canonical Ruby interpreter is.

Currently, the tree walker and vm are implemented
side-by-side. By default the tree walking evaluator
will be used. If Nitrogen is ran with the -compile
flag, the vm will be used. Both consume the same
syntax tree so they don't interfere.

Once everything gets a bit cleaner and more thought
out, documentation will be written for the opcodes
and how the vm works.