Implemented everything except classes and exceptions
There's still a couple bugs with closures, but looks
and if statements should be ok. I need to do some more
thorough testing with deeply nested statments, but
it seems ok.

All tests that don't use special builtins or exceptions
or classes are passing. examples/object.ni is failing
I need to look at it.