Started docs for VM, made SCGI server use VM with -compile
The webserver script example have fixed for loop iterators.
The SCGI will use the compiler and VM if the server is
started with `-compile`.

Started documentation on the VM and bytecodes. It needs
to be expanded. I realy like the JVM documentation as
a template for this. Show the arguments and stack manipulation
very plainly.