Implements basic classes in the VM
This commit adds basic support for classes in the VM.
This has not been tested extensivly nor have
interpreter-land classes been implemented yet. Classes
don't support inhertence yet. The groundwork for it should
be in place, but there's still a few final touches that
need to happen.

This commit also changes a fundamental behaviour with
classes. Previously, object properties could be used
with calling "this". That is no longer the case. To access
object properties, "this" must be used.

Added 4 opcodes:

- BuildClass
- MakeInstance
- LoadAttribute
- StoreAttribute