Closes #16 - Consolidate all external include functions
include, require, and module are now a single import function.
The function will load a file as is appropiate for the situation.
It also supposed .nib files now. Apparently I didn't add that before.

Also fixed a couple issues with try/catch blocks and empty or
essentially empty catch blocks. If the stack is empty, slap
on a Null.