Issue #24 - Write compiled forms of source scripts on import
When a script is imported, a .nib version is checked before
a .ni. If the nib exists, its checked against the source file
for correct mod time. If the modtime is wrong, or the nib doesn't
exist, the source file is compiled and then written to a nib file
for future use. This is the same as Python's original caching
strategy before using __pycache__. I may go that route later
or even something completely different, but for now this works.