Fixed odd bug causing exhausted stack
For some reason the frame stack was being exhausted
during some method executions. The generated bytecode
was correct and the stack size was generated correctly.
The best I could determine is the called function's
return value was not being properly pushed onto the callee's
stack. Putting the function call on a separate line fixes
it but I don't know why. It should have the same effect as
the original line.