Add line numbers to code blocks
Line numbers now appear when bytecode is disassembled
and when a stack trace is generated. Line numbers are
generated by marking the code offset and line number
where a line starts. This way, each bytecode doesn't
need a line number. Only when the line changes does
it need to be marked.

I originally wanted to use the same process that Python
uses, but since my opcodes are variable length it's
more difficult to figure out bytecode offsets based
on the program counter. In Python, every bytecode is
two bytes so the index is simply pc/2. But with variable
opcodes, the entire code stream needs to be stepped over
to figure out a bytecode's index. The full offset
method takes up more space, but it's not too bad. And
I consider the advantage to be well worth it.