Implement interfaces and implements operator
Interfaces are nothing more than a collection of methods and the number
of parameters each takes. A class, instance, or interface implements
an interface if the object has methods with the same name and number
of parameters as the interface. The object may define more methods.

The "implements" keyword returns a boolean value if an object implements
and interface. The interface to check for is on the right hand side
of the expression.

OBJECT implements INTERFACE