Added match expression for simple conditionals
The match expression is sugar over the if/elif/else structure. The match
expression is only known at the parser where it's transformed into an
IfExpression AST node. This way the rest of the compiler doesn't need to
know about them.

Eventually the cases could be more complex by adding support for range
checks or other conditionals. Since this is just a transformed if
statement, anything can be allowed.