Move vm and related packages to a single package
The idea is that the VM can be in an independent
single package and used with multiple frontend
compilers and languages. Right now, the elemental
package still relies on the ast package. The AST
probably also belongs in the elemental package,
but I'm not sure yet. There also needs to be
mechanism that de-couples module imports from
the ast package. This could probably be done by
giving the VM an import function to be called
whenever an IMPORT opcode is executed. This
would hand off the actualy importing to the support
application using the VM.